requirements.txt
opencv-contrib-python==4.6.0.66
pyserial==3.5
sympy==1.13.2